requirements.txt
frappe
qrcode
pillow
six